improving code

+++ app/src/main/java/com/gcalvocr/notes/ui/notes/adapters/NoteListAdapter.kt
@@ -21,7 +21,7 @@ class NoteListAdapter(
         notifyDataSetChanged()
     }
 
-    fun removeData(id: Int){
+    fun removeData(id: Long){
         val index = data.indexOfFirst { item -> item.id == id }
         data.removeAt(index)
         notifyDataSetChanged()
